backend/requirements.txt
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
Django==6.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-extensions==4.1
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-decouple==3.8
redis==7.4.0
requests==2.33.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
watchfiles==1.1.1
whitenoise==6.12.0
zope.interface==8.2
